// apps/white-label/src/types/database.ts
export interface Page {
  id: string;
  company_id: string;
  company_database_id?: string;
  name: string;
  meta_data?: Record<string, any>;
  created_at: string;
  updated_at: string;
}